Moving extractConcreteTypes to PropMeta
